+++ cli/cli/kurtosis_config/config_version_migrators.go
@@ -1,1 +1,0 @@
-package kurtosis_config