+++ cli/cli/kurtosis_config/kurtosis_config_store.go
@@ -4,10 +4,10 @@ import (
 	"github.com/go-yaml/yaml"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/host_machine_directories"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_config/config_version"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_config/overrides_objects/v0"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_config/overrides_objects/v1"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_config/overrides_objects/v2"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_config/resolved_config"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_config/v0"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_config/v1"
-	v2 "github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_config/v2"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"io/ioutil"
@@ -249,64 +249,3 @@ func migrateConfigOverridesToLatest(configFileBytes []byte) (*v1.KurtosisConfigV
 	return resultConfig, nil
 }
 
-func migrateV0ToV1(uncastedConfig interface{}) (*v1.KurtosisConfigV1, error) {
-	// cast "uncastedConfig" to current version we're upgrading from
-	castedOldConfig, ok := uncastedConfig.(*v0.KurtosisConfigV0)
-	if !ok {
-		return nil, stacktrace.NewError(
-			"Failed to cast old configuration '%+v' to expected configuration struct",
-			uncastedConfig,
-		)
-	}
-	// create a new configuration object to represent the migrated work
-	newConfig := &v1.KurtosisConfigV1{
-		ConfigVersion:     config_version.ConfigVersion_v1,
-		ShouldSendMetrics: castedOldConfig.ShouldSendMetrics,
-		KurtosisClusters:  nil,
-	}
-	return newConfig, nil
-}
-
-func migrateV1ToV2(uncastedConfig interface{}) (*v2.KurtosisConfigV2, error) {
-	// cast "uncastedConfig" to current version we're upgrading from
-	castedOldConfig, ok := uncastedConfig.(*v1.KurtosisConfigV1)
-	if !ok {
-		return nil, stacktrace.NewError(
-			"Failed to cast old configuration '%+v' to expected configuration struct",
-			uncastedConfig,
-		)
-	}
-
-	// Migrate cluster configs across
-	var newClusters map[string]*v2.KurtosisClusterConfigV2
-	if castedOldConfig.KurtosisClusters != nil {
-		newClusters = map[string]*v2.KurtosisClusterConfigV2{}
-		for oldClusterName, oldClusterConfig := range *castedOldConfig.KurtosisClusters {
-			oldKubernetesConfig := oldClusterConfig.Config
-
-			var newKubernetesConfig *v2.KubernetesClusterConfigV2
-			if oldKubernetesConfig != nil {
-				newKubernetesConfig = &v2.KubernetesClusterConfigV2{
-					KubernetesClusterName:  oldKubernetesConfig.KubernetesClusterName,
-					StorageClass:           oldKubernetesConfig.StorageClass,
-					EnclaveSizeInMegabytes: oldKubernetesConfig.EnclaveSizeInMegabytes,
-				}
-			}
-
-			newClusterConfig := &v2.KurtosisClusterConfigV2{
-				Type:   oldClusterConfig.Type,
-				Config: newKubernetesConfig,
-			}
-			newClusters[oldClusterName] = newClusterConfig
-		}
-	}
-
-	// create a new configuration object to represent the migrated work
-	newConfig := &v2.KurtosisConfigV2{
-		ConfigVersion:     config_version.ConfigVersion_v2,
-		ShouldSendMetrics: castedOldConfig.ShouldSendMetrics,
-		KurtosisClusters:  newClusters,
-	}
-
-	return newConfig, nil
-}