+++ cli/cli/kurtosis_config/overrides_objects/v1/kurtosis_config_v1.go
@@ -2,6 +2,15 @@ package v1
 
 import "github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_config/config_version"
 
+/*
+!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+                           DO NOT CHANGE THIS FILE!
+  If you change this file, it will break config for users who have instantiated an
+           overrides file with this version of config overrides!
+    Instead, to make changes, you will need to add a new version of the config
+!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+*/
+
 // NOTE: All new YAML property names here should be kebab-case because
 //a) it's easier to read b) it's easier to write
 //c) it's consistent with previous properties and changing the format of