+++ cli/cli/kurtosis_config/v2/kubernetes_cluster_config_v2.go
@@ -1,8 +1,0 @@
-package v2
-
-type KubernetesClusterConfigV2 struct {
-	KubernetesClusterName *string `yaml:"kubernetes-cluster-name,omitempty"`
-	StorageClass *string `yaml:"storage-class,omitempty"`
-	EnclaveSizeInMegabytes *uint `yaml:"enclave-size-in-megabytes,omitempty"`
-}
-